**Changed** This package goal is to extend the functionality of Addressables and now is in parity with that goal - Added the dependency to Unity's Addressables package - Added *AddressablesAssetLoader* to main runtime - Added *AddressablesIdsGenerator* to editor part of the package
