+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-08-08
+
+**Changed**
+This package goal is to extend the functionality of Addressables and now is in parity with that goal
+- Added the dependency to Unity's Addressables package
+- Added *AddressablesAssetLoader* to main runtime
+- Added *AddressablesIdsGenerator* to editor part of the package
+
 ## [0.4.0] - 2020-07-26
 
 - Added *AddressablesLoader* to the samples package with the asset loader implementation for the asset loader
