+++ Editor/GameLovers.AssetLoader.Editor.asmdef
@@ -1,10 +1,12 @@
 {
-    "name": "GameLovers.AssetLoader",
+    "name": "GameLovers.AssetLoader.Editor",
     "references": [
-        "GUID:9e24947de15b9834991c9d8411ea37cf",
-        "GUID:84651a3751eca9349aac36a66bba901b"
+        "GUID:69448af7b92c7f342b298e06a37122aa",
+        "GUID:9e24947de15b9834991c9d8411ea37cf"
     ],
-    "includePlatforms": [],
+    "includePlatforms": [
+        "Editor"
+    ],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
