+++ Runtime/AddressablesAssetLoader.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 773498f63675494c826ec8b85e5a9128
+timeCreated: 1596907155
