--- Samples~/AddressablesLoader/AddressablesLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0974504c9b7a14d79860d72f62469075
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
