--- Samples~/AddressablesLoader/Editor.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a5739548c2a94604bed8ffc51ac7cc4d
-timeCreated: 1595892245
