+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "com.gamelovers.assetloader",
   "displayName": "Asset Loader",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "unity": "2019.3",
   "description": "This package extends the loading functionality of assets from Unity Addressables with the possibility to generate new info about each asset addressable.\n\nThe asset info generator can be executed in Tools > Generate AddressableIds",
   "type": "library",
